refactor: rename feedback handling utility and streamline handleSubmit function
